REFACTOR :: 로그인 화면
- 간단한 수정
