+++ src/components/Login/Login_hide/Login_hide.scss
@@ -1,13 +1,3 @@
-.order-inquiry-div {
-  text-align: center;
-  margin-top: 20px;
-}
-
-.order-inquiry-button {
-  background-color: white;
-  font-size: 17px;
-}
-
 .ordernum-inquiry-div {
   text-align: center;
   margin-top: 20px;
@@ -40,15 +30,4 @@
   height: 50px;
   text-indent: 15px;
   font-size: 15px;
-}
-
-.copyright-div {
-  p {
-    color: gray;
-    font-size: 10px;
-  }
-  text-align: center;
-  position: fixed;
-  left: 930px;
-  bottom: 30px;
 }
