+++ src/pages/Login/Login.js
@@ -1,11 +1,10 @@
-import Login_main from "../../components/Login_main/Login_main";
-import Login_hide from "../../components/Login_hide/Login_hide";
+import Login_main from "../../components/Login/Login_main/Login_main";
+import Login_hide from "../../components/Login/Login_hide/Login_hide";
 
 export default function Login() {
   return (
     <>
       <Login_main />
-      <Login_hide />
     </>
   );
 }
